Adequando PEP8

+++ testes/teste12.py
@@ -27,9 +27,9 @@
 #
 # 6- Observar a saída exibida no terminal.
 
-from inewave.newave.pmo import LeituraPMO
-from inewave.newave.parp import LeituraPARp
-from inewave.config import REES
+from inewave.newave.pmo import LeituraPMO  # type: ignore
+from inewave.newave.parp import LeituraPARp  # type: ignore
+from inewave.config import REES  # type: ignore
 from typing import Dict
 import numpy as np
 from parpa.yulewalker import YuleWalkerPARA
@@ -58,10 +58,10 @@ periodo_max_dif_ree: Dict[int, int] = {ree: 0
                                        for ree in IDS_REES}
 ordem_max_dif_ree: Dict[int, int] = {ree: 0
                                      for ree in IDS_REES}
-coef_o_max_dif_ree: Dict[int, int] = {ree: 0
-                                      for ree in IDS_REES}
-coef_e_max_dif_ree: Dict[int, int] = {ree: 0
-                                      for ree in IDS_REES}
+coef_o_max_dif_ree: Dict[int, float] = {ree: 0
+                                        for ree in IDS_REES}
+coef_e_max_dif_ree: Dict[int, float] = {ree: 0
+                                        for ree in IDS_REES}
 # Máxima diferença percentual por REE
 max_dif_percent_ree: Dict[int, float] = {ree: -1e4
                                          for ree in IDS_REES}
@@ -71,10 +71,10 @@ periodo_max_dif_perc_ree: Dict[int, int] = {ree: 0
                                             for ree in IDS_REES}
 ordem_max_dif_perc_ree: Dict[int, int] = {ree: 0
                                           for ree in IDS_REES}
-coef_o_max_dif_perc_ree: Dict[int, int] = {ree: 0
-                                           for ree in IDS_REES}
-coef_e_max_dif_perc_ree: Dict[int, int] = {ree: 0
-                                           for ree in IDS_REES}
+coef_o_max_dif_perc_ree: Dict[int, float] = {ree: 0
+                                             for ree in IDS_REES}
+coef_e_max_dif_perc_ree: Dict[int, float] = {ree: 0
+                                             for ree in IDS_REES}
 
 
 # Calcula as autocorrelações parciais e compara com o arquivo
@@ -107,20 +107,20 @@ for ree in IDS_REES:
                 dif = abs(c - oficial)
                 dif_percentual = 100 * abs(c - oficial) / oficial
                 if dif > max_dif_ree[ree]:
-                        max_dif_ree[ree] = dif
-                        max_dif_percent_ree[ree] = dif_percentual
-                        ano_max_dif_ree[ree] = ano
-                        periodo_max_dif_ree[ree] = mes
-                        ordem_max_dif_ree[ree] = i
-                        coef_e_max_dif_ree[ree] = c
-                        coef_o_max_dif_ree[ree] = oficial
+                    max_dif_ree[ree] = dif
+                    max_dif_percent_ree[ree] = dif_percentual
+                    ano_max_dif_ree[ree] = ano
+                    periodo_max_dif_ree[ree] = mes
+                    ordem_max_dif_ree[ree] = i
+                    coef_e_max_dif_ree[ree] = c
+                    coef_o_max_dif_ree[ree] = oficial
                 if dif_percentual > max_dif_percent_ree[ree]:
-                        max_dif_percent_ree[ree] = dif_percentual
-                        ano_max_dif_perc_ree[ree] = ano
-                        periodo_max_dif_perc_ree[ree] = mes
-                        ordem_max_dif_perc_ree[ree] = i
-                        coef_e_max_dif_perc_ree[ree] = c
-                        coef_o_max_dif_perc_ree[ree] = oficial
+                    max_dif_percent_ree[ree] = dif_percentual
+                    ano_max_dif_perc_ree[ree] = ano
+                    periodo_max_dif_perc_ree[ree] = mes
+                    ordem_max_dif_perc_ree[ree] = i
+                    coef_e_max_dif_perc_ree[ree] = c
+                    coef_o_max_dif_perc_ree[ree] = oficial
             mes += 1
 
 print("")
